Use LVname not VG-LVname when adding LVs
The VG-LVname string is not a key in the dictionary constructed from the 'lvm
lvs' output and if by some chance (really high on Python 3) e.g. a thin pool is
requested to be added by a thin LV, using VG-LVname as the key raises a KeyError
while the info about the pool LV is in the dictionary as expected.