Use the storaged's iSCSI DBus API instead of libiscsi
libiscsi is, at this point, a giant hack and its python bindings are another
giant hack on top of that hack. Instead of dealing with these, we should just
use the DBus API for iSCSI provided by the storaged daemon.

The only change in behaviour should be that all nodes are marked for automatic
startup while before this patch it was only nodes that weren't required for the
root file system marked that way. Is there a problem with marking nodes required
for the root file system (and thus started up by dracut)? If so, isn't that a
bug in openiscsi?