Define types for internal LVs and their type-specific methods and props
Internal LVs may be of multiple types which determine what can and cannot be
done with them. E.g. an internal metadata LV can be resized, but others
cannot. Also some of the LVs take extra space in the VG (not included in their
parent LV's size) and some don't.