Adapt to using segment types more widely
We now use the LV segment type to determine a lot of things. But working with
right and more realistic values of LV.seg_type means we cannot rely on it having
values we would like. E.g. cached LVs have a segment type 'cache' no matter if
they (the data) are in fact linear, striped or whatever.